+++ app/src/main/java/com/example/studenttaskmanagement/model/Category.java
@@ -1,24 +1,24 @@
 package com.example.studenttaskmanagement.model;
 
 public class Category {
-    private int id;
+    private long id;
     private String name;
     private String color;
 
     public Category() {
     }
 
-    public Category(int id, String name, String color) {
+    public Category(long id, String name, String color) {
         this.id = id;
         this.name = name;
         this.color = color;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
